RF FreestyleProjectTest classes
